+++ runtime/Java/src/org/antlr/v4/runtime/BaseErrorListener.java
@@ -22,7 +22,7 @@ public class BaseErrorListener implements ANTLRErrorListener {
 	public void syntaxError(Recognizer<?, ?> recognizer,
 							Object offendingSymbol,
 							int line,
-							int charPositionInLine,
+							long charPositionInLine,
 							String msg,
 							RecognitionException e)
 	{
@@ -31,8 +31,8 @@ public class BaseErrorListener implements ANTLRErrorListener {
 	@Override
 	public void reportAmbiguity(Parser recognizer,
 								DFA dfa,
-								int startIndex,
-								int stopIndex,
+								long startIndex,
+								long stopIndex,
 								boolean exact,
 								BitSet ambigAlts,
 								ATNConfigSet configs)
@@ -42,8 +42,8 @@ public class BaseErrorListener implements ANTLRErrorListener {
 	@Override
 	public void reportAttemptingFullContext(Parser recognizer,
 											DFA dfa,
-											int startIndex,
-											int stopIndex,
+											long startIndex,
+											long stopIndex,
 											BitSet conflictingAlts,
 											ATNConfigSet configs)
 	{
@@ -52,8 +52,8 @@ public class BaseErrorListener implements ANTLRErrorListener {
 	@Override
 	public void reportContextSensitivity(Parser recognizer,
 										 DFA dfa,
-										 int startIndex,
-										 int stopIndex,
+										 long startIndex,
+										 long stopIndex,
 										 int prediction,
 										 ATNConfigSet configs)
 	{
