+++ runtime/Java/src/org/antlr/v4/runtime/ProxyErrorListener.java
@@ -33,7 +33,7 @@ public class ProxyErrorListener implements ANTLRErrorListener {
 	public void syntaxError(Recognizer<?, ?> recognizer,
 							Object offendingSymbol,
 							int line,
-							int charPositionInLine,
+							long charPositionInLine,
 							String msg,
 							RecognitionException e)
 	{
@@ -45,8 +45,8 @@ public class ProxyErrorListener implements ANTLRErrorListener {
 	@Override
 	public void reportAmbiguity(Parser recognizer,
 								DFA dfa,
-								int startIndex,
-								int stopIndex,
+								long startIndex,
+								long stopIndex,
 								boolean exact,
 								BitSet ambigAlts,
 								ATNConfigSet configs)
@@ -59,8 +59,8 @@ public class ProxyErrorListener implements ANTLRErrorListener {
 	@Override
 	public void reportAttemptingFullContext(Parser recognizer,
 											DFA dfa,
-											int startIndex,
-											int stopIndex,
+											long startIndex,
+											long stopIndex,
 											BitSet conflictingAlts,
 											ATNConfigSet configs)
 	{
@@ -72,8 +72,8 @@ public class ProxyErrorListener implements ANTLRErrorListener {
 	@Override
 	public void reportContextSensitivity(Parser recognizer,
 										 DFA dfa,
-										 int startIndex,
-										 int stopIndex,
+										 long startIndex,
+										 long stopIndex,
 										 int prediction,
 										 ATNConfigSet configs)
 	{
