+++ runtime/Java/src/org/antlr/v4/runtime/Recognizer.java
@@ -172,8 +172,8 @@ public abstract class Recognizer<Symbol, ATNInterpreter extends ATNSimulator> {
 	/** What is the error header, normally line/character position information? */
 	public String getErrorHeader(RecognitionException e) {
 		int line = e.getOffendingToken().getLine();
-		int charPositionInLine = e.getOffendingToken().getCharPositionInLine();
-		return "line "+line+":"+charPositionInLine;
+		long charPositionInLine = e.getOffendingToken().getCharPositionInLine();
+		return "line " + line + ":" + charPositionInLine;
 	}
 
 	/** How should a token be displayed in an error message? The default
