+++ runtime/Java/src/org/antlr/v4/runtime/atn/ContextSensitivityInfo.java
@@ -42,7 +42,7 @@ public class ContextSensitivityInfo extends DecisionEventInfo {
 	 */
 	public ContextSensitivityInfo(int decision,
 								  ATNConfigSet configs,
-								  TokenStream input, int startIndex, int stopIndex)
+								  TokenStream input, long startIndex, long stopIndex)
 	{
 		super( decision, configs, input, startIndex, stopIndex, true);
 	}
